Frontend, BitBoxBase: Rename bitboxbase to bitboxbaseconnect
