+++ frontends/web/src/routes/bitboxbase/bitboxbaseconnect.tsx
